tools/execution_tracer/requirements.txt
pyelftools==0.30
